update all evaluations to have the same format
